Improve support for various dtypes Almost all convolutional layers now support some combination of int32, int64, float16, float32, and float64 inputs. Layers based on Conv now autocast integer adjacency matrices to float if the node features are float Tests for dtype compatibility Improve support for keras.Layer automatic argument passing

+++ spektral/layers/convolutional/gat_conv.py
@@ -155,7 +155,7 @@ class GATConv(Conv):
                 name="bias",
             )
 
-        self.dropout = Dropout(self.dropout_rate)
+        self.dropout = Dropout(self.dropout_rate, dtype=self.dtype)
         self.built = True
 
     def call(self, inputs):
